requirements.txt
python-telegram-bot[job-queue]==21.5
supabase==2.9.1
python-dotenv==1.0.0
fastapi
uvicorn